+++ lib/src/_constants.dart
@@ -1,5 +1,3 @@
-const kApp = #spry.app;
-const kRequest = #spry.event.request;
 const kNext = #spry.event.next;
 const kParams = #spry.event.params;
 const kClientAddress = #spry.event.client_address;