+++ lib/src/utils/create_event.dart
@@ -1,16 +1,16 @@
-import '../_constants.dart';
 import '../http/request.dart';
 import '../types.dart';
 
 /// Creates a new Spry [Event] instance.
-Event createEvent(Spry app, Request request) {
-  return _EventImpl()
-    ..set(kRequest, request)
-    ..set(kApp, app);
+Event createEvent<Raw>(
+    {required Spry app, required Request request, required Raw raw}) {
+  return _EventImpl(app: app, raw: raw, request: request);
 }
 
-class _EventImpl implements Event {
-  final locals = <Object?, Object?>{};
+class _EventImpl<Raw> implements Event {
+  _EventImpl({required this.app, required this.raw, required this.request});
+
+  late final locals = <Object?, Object?>{};
 
   @override
   T? get<T>(Object? key) {
@@ -29,4 +29,13 @@ class _EventImpl implements Event {
   void set<T>(Object? key, T value) {
     locals[key] = value;
   }
+
+  @override
+  final Raw raw;
+
+  @override
+  final Request request;
+
+  @override
+  final Spry app;
 }